Move get files query to @/lib
